pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
name = "DMD_toolbox"
authors = [
    {name = "Gabriel Seigneur", email = "gabseig@gmail.com"},
    {name = "Hugo Hakem", email = "hugo.hakem@berkeley.edu"}
]
description = "Toolbox to setup and use a Digital Micromirror Device EVM from Texas Instruments."
license = { file = "LICENSE.md" }
readme = "README.md"
version = "0.1.0"

requires-python = ">=3.10"
dependencies = [
    "matplotlib",
    "numpy",
    "scipy",
    "pillow",
    "ipykernel",
    "pip",
    "jupyter"
]

[project.optional-dependencies]
docs = [
    "sphinx",
    "myst-parser",
    "sphinx-copybutton",
    "furo",
    "doc8",
    "rstcheck",
    "esbonio",
    "pandoc",
]
lint = ["ruff"]
mypy = ["mypy", "types-PyYAML"]
dev = ["DMD_toolbox[docs, lint, mypy]"]

[tool.hatch.build.targets.wheel]
packages = ["src/dmd_toolbox"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-arm64", "osx-64", "win-64"]

[tool.pixi.pypi-dependencies]
DMD_toolbox = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.feature.docs.dependencies]
pandoc = ">=3.1.11"

[tool.pixi.environments]
default = { solve-group = "default" }
docs = { features = ["docs"], solve-group = "default" }
dev = { features = ["docs"], solve-group = "default" }

[tool.ruff]
line-length = 120
extend-include = ["*.ipynb"]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
ignore_missing_imports = true
explicit_package_bases = true
check_untyped_defs = true
warn_unreachable = true

